// types.ts
export interface ServiceCatalogEndpoint {
  id: string;
  interface: 'public' | 'internal' | 'admin';
  region: string;
  region_id: string;
  url: string;
}

export interface ServiceCatalogEntry {
  endpoints: ServiceCatalogEndpoint[];
  id: string;
  name: string;
  type: string; // e.g., 'compute', 'volumev3', 'network', 'image', 'identity'
}

export interface KeystoneProject {
  id: string;
  name: string;
  domain: { id: string; name: string; };
}

export interface KeystoneUser {
  id: string;
  name: string;
  domain: { id: string; name: string; };
  enabled?: boolean;
  domain_id?: string;
}

export interface Role {
    id: string;
    name: string;
}
// ...
export interface SessionExportData {
  authUrl: string;
  authToken: string;
  serviceCatalog: ServiceCatalogEntry[];
  project?: KeystoneProject | null;
  user?: KeystoneUser | null;
  roles?: Role[] | null;
}